refactor: thresholdViolation data-type

+++ src/main/java/at/qe/skeleton/dtos/ThresholdViolationDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
 public record ThresholdViolationDTO(
         Long id,
         Metric metric,
-        Long value,
+        Float value,
         ViolationStatus violationStatus,
         LocalDateTime endTime,
         LocalDateTime startTime,
